pyproject.toml
[project]
name = "mcp-test-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.4.1",
    "openai==1.66.3",
    "python-dotenv==1.0.1",
]
